Added payment method using iyzico

+++ Bookworm/Bookworm.API/Controllers/OrderController.cs
@@ -2,6 +2,9 @@ using Bookworm.API.Data;
 using Bookworm.API.Dtos;
 using Bookworm.API.Entity;
 using Bookworm.API.Extensions;
+using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -13,12 +16,15 @@ namespace Bookworm.API.Controller;
 public class OrderController: ControllerBase
 {
     private readonly DataContext _context;
+    private readonly IConfiguration _config;
 
-    public OrderController (DataContext context)
+    public OrderController (DataContext context, IConfiguration config)
     {
         _context = context;
+        _config = config;
     }
 
+    #region Get Orders
     [HttpGet]
     [Authorize(Roles = "Customer")]
     public async Task<ActionResult<List<OrderDto>>> GetOrder()
@@ -29,7 +35,9 @@ public class OrderController: ControllerBase
         .Where(i=>i.CustomerId==User.Identity!.Name)
         .ToListAsync();
     }
+    #endregion
 
+    #region Get Orders By Admin
     [HttpGet("get-by-admin")]
     [Authorize(Roles = "Admin")]
     public async Task<ActionResult<List<OrderDto>>> GetOrderByAdmin()
@@ -39,7 +47,9 @@ public class OrderController: ControllerBase
         .OrderToDto()
         .ToListAsync();
     }
+    #endregion
 
+    #region Get Order
     [HttpGet("{id}")]
     [Authorize(Roles = "Customer")]
         public async Task<ActionResult<OrderDto?>> GetOrder(int id)
@@ -50,7 +60,9 @@ public class OrderController: ControllerBase
                         .Where(i => i.CustomerId == User.Identity!.Name && i.Id == id)
                         .FirstOrDefaultAsync();
         }
+    #endregion
 
+    #region Create Order
     [HttpPost]
     [Authorize(Roles = "Customer")]
     public async Task<ActionResult<Order>> CreateOrder(CreateOrderDto createOrderDto)
@@ -63,13 +75,13 @@ public class OrderController: ControllerBase
 
         if(cart==null) return BadRequest(new ProblemDetails{Title="Problem getting cart"});
 
-        var items = new List<OrderItem>();
+        var items = new List<Entity.OrderItem>();
 
         foreach (var item in cart.CartItems)
         {
             var book = await _context.Books.FindAsync(item.BookId);
 
-            var orderItem = new OrderItem
+            var orderItem = new Entity.OrderItem
             {
                 BookId = book!.Id,
                 BookName = book.Title!,
@@ -98,6 +110,16 @@ public class OrderController: ControllerBase
                 DeliveryFree = deliveryFree
             };
 
+            var paymentResult = await ProcessPayment(createOrderDto, cart);
+            
+            if (paymentResult.Status == "failure")
+            {
+                return BadRequest(new ProblemDetails { Title = paymentResult.ErrorMessage });
+            }
+
+            order.ConversationId = paymentResult.ConversationId;
+            order.BasketId = paymentResult.BasketId;
+
             _context.Orders.Add(order);
             _context.Carts.Remove(cart);
 
@@ -108,14 +130,92 @@ public class OrderController: ControllerBase
             
             return BadRequest(new ProblemDetails { Title = "Problem getting order" });
     }
+    #endregion
 
+    #region Process Payment
+    private async Task<Payment> ProcessPayment(CreateOrderDto createOrderDto, Cart cart)
+        {
+            Options options = new Options();
+            options.ApiKey = _config["PaymentAPI:APIKey"];
+            options.SecretKey = _config["PaymentAPI:SecretKey"];
+
+            options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+            CreatePaymentRequest request = new CreatePaymentRequest();
+            request.Locale = Locale.TR.ToString();
+            request.ConversationId = Guid.NewGuid().ToString();
+            request.Price = cart.CalculateTotal().ToString();
+            request.PaidPrice = cart.CalculateTotal().ToString();
+            request.Currency = Currency.TRY.ToString();
+            request.Installment = 1;
+            request.BasketId = cart.CartId.ToString();
+            request.PaymentChannel = PaymentChannel.WEB.ToString();
+            request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+            PaymentCard paymentCard = new PaymentCard();
+            paymentCard.CardHolderName = createOrderDto.CardName;
+            paymentCard.CardNumber = createOrderDto.CardNumber;
+            paymentCard.ExpireMonth = createOrderDto.CardExpireMonth;
+            paymentCard.ExpireYear = createOrderDto.CardExpireYear;
+            paymentCard.Cvc = createOrderDto.CardCvc;
+            paymentCard.RegisterCard = 0;
+            request.PaymentCard = paymentCard;
+
+            Buyer buyer = new Buyer();
+            buyer.Id = "BY789";
+            buyer.Name = createOrderDto.FirstName;
+            buyer.Surname = createOrderDto.LastName;
+            buyer.GsmNumber = createOrderDto.Phone;
+            buyer.Email = "email@email.com";
+            buyer.IdentityNumber = "74300864791";
+            buyer.LastLoginDate = "2015-10-05 12:43:35";
+            buyer.RegistrationDate = "2013-04-21 15:12:09";
+            buyer.RegistrationAddress = createOrderDto.AddresLine;
+            buyer.Ip = "85.34.78.112";
+            buyer.City = createOrderDto.City;
+            buyer.Country = "Türkiye";
+            buyer.ZipCode = "34732";
+            request.Buyer = buyer;
+
+            Address shippingAddress = new Address();
+            shippingAddress.ContactName = createOrderDto.FirstName + " " + createOrderDto.LastName;
+            shippingAddress.City = createOrderDto.City;
+            shippingAddress.Country = "Türkiye";
+            shippingAddress.Description = createOrderDto.AddresLine;
+            shippingAddress.ZipCode = "34742";
+
+            request.ShippingAddress = shippingAddress;
+            request.BillingAddress = shippingAddress;
+
+            List<BasketItem> basketItems = new List<BasketItem>();
+
+            foreach (var item in cart.CartItems)
+            {
+                BasketItem basketItem = new BasketItem();
+                basketItem.Id = item.BookId.ToString();
+                basketItem.Name = item.Book.Title;
+                basketItem.Category1 = "Saat";
+                basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+                basketItem.Price = ((double)item.Book.Price * item.Quantity).ToString();
+                basketItems.Add(basketItem);
+            }
+
+            request.BasketItems = basketItems;
+
+            return await Payment.Create(request, options);
+        }
+    #endregion
+
+    #region Get Total Order Count
     [HttpGet("count")]
     [Authorize(Roles = "Admin")]
     public async Task<ActionResult<int>> GetTotalOrderCount()
     {
         return await _context.Orders.CountAsync();
     }
+    #endregion
 
+    #region Get Today Orders
     [HttpGet("today")]
     [Authorize(Roles = "Admin")]
     public async Task<ActionResult<List<OrderDto>>> GetTodayOrders()
@@ -127,7 +227,9 @@ public class OrderController: ControllerBase
             .OrderToDto()
             .ToListAsync();
     }
+    #endregion
 
+    #region Get Pending Orders
     [HttpGet("pending-orders")]
     [Authorize(Roles = "Admin")]
     public async Task<ActionResult<List<OrderDto>>> GetPendingOrders()
@@ -137,4 +239,5 @@ public class OrderController: ControllerBase
             .OrderToDto()
             .ToListAsync();
     }
+    #endregion
 }
